refactor(adhoc): extract picker modal and empty-state to sibling file
AdHocLoggerView.tsx was over the 1000-line cap after the plan-editor
wiring. The picker modal and tip-style empty state had no business
sitting at the bottom of the view file — they're self-contained,
prop-driven, and independently understandable. Moved them to
components/adhoc/AdHocExercisePickerModal.tsx along with the PickerMode
type. View is now ~900 lines and the picker has a clearer home.
