+++ components/adhoc/AdHocLoggerView.tsx
@@ -1,37 +1,28 @@
 'use client'
 
-import { Plus, RefreshCw, Sparkles, Trash2 } from 'lucide-react'
+import { Plus, RefreshCw, Trash2 } from 'lucide-react'
 import { useRouter } from 'next/navigation'
 import { useCallback, useEffect, useRef, useState } from 'react'
 import { createPortal } from 'react-dom'
-import {
-  type ExerciseDefinition,
-  ExerciseSearchInterface,
-} from '@/components/exercise-selection/ExerciseSearchInterface'
-import ExerciseDefinitionEditorModal from '@/components/features/exercise-definition/ExerciseDefinitionEditorModal'
+import type { ExerciseDefinition } from '@/components/exercise-selection/ExerciseSearchInterface'
 import { WorkoutRollupModal } from '@/components/features/training/WorkoutRollupModal'
 import { useToast } from '@/components/ToastProvider'
 import { Button } from '@/components/ui/Button'
 import { LoadingFrog } from '@/components/ui/loading-frog'
-import {
-  Dialog,
-  DialogBody,
-  DialogContent,
-  DialogDescription,
-  DialogFooter,
-  DialogHeader,
-  DialogTitle,
-} from '@/components/ui/radix/dialog'
-import { TipAnnotation } from '@/components/ui/TipAnnotation'
 import ExerciseActionsFooter from '@/components/workout-logging/ExerciseActionsFooter'
 import ExerciseDisplayTabs from '@/components/workout-logging/ExerciseDisplayTabs'
 import ExerciseLoggingHeader from '@/components/workout-logging/ExerciseLoggingHeader'
 import type { QuickAction } from '@/components/workout-logging/ExerciseQuickActionsMenu'
 import ExitWorkoutConfirm from '@/components/workout-logging/ExitWorkoutConfirm'
-import WorkoutPlanEditor from '@/components/workout-logging/WorkoutPlanEditor'
 import SetLoggingForm, {
   type ExpandedInput,
 } from '@/components/workout-logging/SetLoggingForm'
+import WorkoutPlanEditor from '@/components/workout-logging/WorkoutPlanEditor'
+import {
+  AdHocEmptyState,
+  AdHocExercisePickerModal,
+  type PickerMode,
+} from './AdHocExercisePickerModal'
 import { useIntensityAccess } from '@/hooks/useIntensityAccess'
 import { useSwipeNavigation } from '@/hooks/useSwipeNavigation'
 import {
@@ -785,7 +776,7 @@ export default function AdHocLoggerView({
               ] satisfies QuickAction[]}
             />
           ) : (
-            <EmptyState />
+            <AdHocEmptyState />
           )}
         </div>
 
@@ -829,7 +820,7 @@ export default function AdHocLoggerView({
       </div>
 
       {pickerMode && (
-        <ExercisePickerModal
+        <AdHocExercisePickerModal
           mode={pickerMode}
           onClose={() => setPickerMode(null)}
           onConfirm={pickerMode.kind === 'add' ? handleAddExercises : handleSwapExercise}
@@ -906,151 +897,5 @@ export default function AdHocLoggerView({
         onClose={handleRollupClose}
       />
     </>
-  )
-}
-
-function EmptyState() {
-  return (
-    <div className="flex-1 flex items-center justify-center px-6 py-8">
-      <TipAnnotation
-        icon={<Sparkles aria-hidden="true" size={20} strokeWidth={1.8} />}
-      >
-        <span className="text-xl sm:text-2xl leading-relaxed text-foreground">
-          Pick your first exercise to start logging. Add as many as you want as you go.
-        </span>
-      </TipAnnotation>
-    </div>
-  )
-}
-
-type PickerMode =
-  | { kind: 'add' }
-  | { kind: 'swap'; replacingName: string; targetId?: string }
-
-function ExercisePickerModal({
-  mode,
-  onClose,
-  onConfirm,
-  isBusy,
-}: {
-  mode: PickerMode
-  onClose: () => void
-  onConfirm: (defs: ExerciseDefinition[]) => void
-  isBusy: boolean
-}) {
-  const isAdd = mode.kind === 'add'
-  // Multi-select state used only in add mode.
-  const [selectedDefs, setSelectedDefs] = useState<ExerciseDefinition[]>([])
-  const selectedIds = new Set(selectedDefs.map((d) => d.id))
-  const [showCreateModal, setShowCreateModal] = useState(false)
-
-  const handleAddToggle = useCallback((def: ExerciseDefinition) => {
-    setSelectedDefs((prev) =>
-      prev.some((d) => d.id === def.id)
-        ? prev.filter((d) => d.id !== def.id)
-        : [...prev, def]
-    )
-  }, [])
-
-  const handleSwapSelect = useCallback(
-    (def: ExerciseDefinition) => {
-      if (isBusy) return
-      onConfirm([def])
-    },
-    [isBusy, onConfirm]
-  )
-
-  const count = selectedDefs.length
-  const title = isAdd ? 'Search Exercises' : 'Swap Exercise'
-  const description = isAdd
-    ? 'Pick one or more to add to your workout'
-    : `Pick a replacement for ${mode.kind === 'swap' ? mode.replacingName : ''}`
-
-  return (
-    <Dialog open onOpenChange={(open) => { if (!open) onClose() }}>
-      <DialogContent
-        showClose={false}
-        fullScreenMobile={true}
-        // Body renders ExerciseSearchInterface; prevent default Radix auto-focus
-        // so the mobile keyboard doesn't pop up automatically (issue #846).
-        onOpenAutoFocus={(e) => e.preventDefault()}
-        className="w-full h-full sm:w-[90vw] sm:max-w-3xl sm:h-auto sm:max-h-[85vh] rounded-none sm:rounded-none border border-border bg-card"
-      >
-        <DialogHeader className="border-b border-border bg-primary py-2">
-          <div className="flex items-center justify-between">
-            <div className="flex-1">
-              <DialogTitle className="text-lg font-bold text-primary-foreground tracking-wider uppercase">
-                {title}
-              </DialogTitle>
-              <DialogDescription className="text-base font-bold text-primary-foreground/70 uppercase tracking-wide">
-                {description}
-              </DialogDescription>
-            </div>
-          </div>
-        </DialogHeader>
-
-        <DialogBody className="flex-1 min-h-0">
-          <ExerciseSearchInterface
-            onExerciseSelect={isAdd ? handleAddToggle : handleSwapSelect}
-            selectedIds={isAdd ? selectedIds : undefined}
-            preloadExercises
-          />
-        </DialogBody>
-
-        <DialogFooter className="border-t border-border bg-card py-2">
-          <div className="flex items-center justify-end gap-2 w-full">
-            <Button variant="secondary" onClick={onClose} doom disabled={isBusy}>
-              Cancel
-            </Button>
-            {isAdd && (
-              <>
-                <Button
-                  variant="secondary"
-                  onClick={() => setShowCreateModal(true)}
-                  doom
-                  disabled={isBusy}
-                >
-                  + Create New Exercise
-                </Button>
-                <Button
-                  variant="primary"
-                  onClick={() => onConfirm(selectedDefs)}
-                  disabled={count === 0 || isBusy}
-                  loading={isBusy}
-                  doom
-                >
-                  {count === 0 ? 'Add' : count === 1 ? 'Add 1 exercise' : `Add all (${count})`}
-                </Button>
-              </>
-            )}
-          </div>
-        </DialogFooter>
-      </DialogContent>
-      <ExerciseDefinitionEditorModal
-        isOpen={showCreateModal}
-        onClose={() => setShowCreateModal(false)}
-        mode="create"
-        onSuccess={(newExercise) => {
-          setShowCreateModal(false)
-          // Auto-select the newly created exercise so the user can add it
-          // straight away without having to re-find it in search results.
-          const def: ExerciseDefinition = {
-            id: newExercise.id,
-            name: newExercise.name,
-            primaryFAUs: newExercise.primaryFAUs,
-            secondaryFAUs: newExercise.secondaryFAUs,
-            equipment: newExercise.equipment,
-            instructions: newExercise.instructions,
-          }
-          if (isAdd) {
-            setSelectedDefs((prev) =>
-              prev.some((d) => d.id === def.id) ? prev : [...prev, def]
-            )
-          } else {
-            onConfirm([def])
-          }
-        }}
-      />
-    </Dialog>
   )
 }
